Add extensive testing for rounding (see is_correct_rounded); fix a bug thus revealed
`is_correct_rounded` manually checks the rounding condition: that the
true value lies in the region

  [posit.prior(), posit.next()]

but is closer to `posit` than to either of them, or the same distance
but `posit` is even.

As in other parts, we check this exhaustively where we can (about ≤p16)
and with proptest where checking all cases is too expensive.

The bug had to do with the corner case where the regime pushes all
exponent and fraction bits out.