First implementation of Posit::round_from(quire)
Now we have: `Quire::from(posit)`, `quire += posit`, and
`Posit::round_from(quire)`, so we have all the pieces to do a basic
summation using the quire! This is illustrated in the updated example
in the crate root doc.

However: this is a first draft implementation. Although correct it is
messy and not optimised enough. Will take this as a perf baseline and
use benchmarks to guide the cleanup and optimisations.